app/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.3
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
python-multipart==0.0.9
jinja2==3.1.4
MarkupSafe==2.1.5
gunicorn==21.2.0
pytest==8.2.1
httpx==0.27.2
